requirements.txt
Pillow
hachoir
pillow-heif
